+++ R/model.R
@@ -3,25 +3,24 @@
 
 #' Train a multi-view model for a single target
 #'
-#' Trains individual models for each view. Each view is modeled by the 
+#' Trains individual models for each view. Each view is modeled by the
 #' function supplied via \code{model.function} which is passed down by
 #' \code{run_misty()}.
 #'
 #' The default model is based on \code{\link[ranger]{ranger}()} for training the
-#' view-specific models with the following parameters: \code{num.trees = 100}, 
+#' view-specific models with the following parameters: \code{num.trees = 100},
 #' \code{importance = "impurity"}, \code{num.threads = 1}, \code{seed = seed}.
 #'
 #' @inheritParams run_misty
 #'
 #' @param target name of the target marker to train models for.
 #'
-#' @return A list containing the trained meta-model, the view-specific 
+#' @return A list containing the trained meta-model, the view-specific
 #' importances and performance estimates.
 #'
 #' @noRd
 build_model <- function(views, target, model.function, model.name,
                         cv.folds, bypass.intra, seed, cached, ...) {
-  
   cache.location <- R.utils::getAbsolutePath(paste0(
     ".misty.temp", .Platform$file.sep,
     views[["misty.uniqueid"]]
@@ -31,7 +30,7 @@ build_model <- function(views, target, model.function, model.name,
     dir.create(cache.location, recursive = TRUE, showWarnings = TRUE)
   }
 
-  expr <- views[["intraview"]][["data"]]
+  expr <- views[["intraview"]]
 
   target.vector <- expr %>% dplyr::pull(target)
 
@@ -39,35 +38,36 @@ build_model <- function(views, target, model.function, model.name,
   ellipsis.args.text <- paste(names(ellipsis.args), ellipsis.args,
     sep = ".", collapse = "."
   )
-  
+
   # returns a list of models
   model.views <- views %>%
     rlist::list.remove(c("misty.uniqueid")) %>%
-    purrr::map(function(view) {
-      
+    purrr::imap(function(view, view.name) {
       model.view.cache.file <-
         paste0(
           cache.location, .Platform$file.sep,
-          "model.", model.name, ".", view[["abbrev"]], ".", target,
-          ".par.", cv.folds, ".", 
+          "model.", model.name, ".", view.name, ".", target,
+          ".par.", cv.folds, ".",
           ellipsis.args.text, ".rds"
         )
 
       if (file.exists(model.view.cache.file) & cached) {
         model.view <- readr::read_rds(model.view.cache.file)
       } else {
-        if ((view[["abbrev"]] == "intra") & bypass.intra) {
+        if ((view.name == "intraview") & bypass.intra) {
           transformed.view.data <-
             tibble::tibble(!!target := target.vector, ".novar" := 1)
         } else {
-          transformed.view.data <- view[["data"]] %>%
+          transformed.view.data <- view %>%
             dplyr::mutate(!!target := target.vector)
         }
-        
+
         # Build model
-        model.view <- model.function(view_data = transformed.view.data,
-                                     target = target, seed = seed, ...)
-        
+        model.view <- model.function(
+          view_data = transformed.view.data,
+          target = target, seed = seed, ...
+        )
+
         if (cached) {
           readr::write_rds(model.view, model.view.cache.file)
         }